Introduce Elm architecture
The `main` function is no longer static but instead uses `Html.program`.
The `Html.program` function takes in record with three vallues: an
initial model, a `view` function, and an `update` function. Note that
functions can be passed as arguments to other functions.

The `update` function takes in an event and the current model and returns
a new model. In this simplest implementation, we're ignoring the event
and always returning the current model. Because the view only changes if
the model changes, we still effectively have a static app.