React to open and close events
We add a `Msg` type that defines two events `OpenSelect` and
`CloseSelect`. The `update` handles both of these by correctly changing
the model. The view gets re-rendered whenever the model changes.

We emit the events from the views in response to a click event handler.

Note that the type signatures for the view functions changed from `Html
a` to `Html Maybe`. That's because the views are now emitting `Msg`
events.