Add the concept of a "selected" value
The whole point of this is to be able to select values. Since it's
possible for no values to be selected, we wrap it in `Maybe`.

The view gets trickier because there are now 4 ways to render the
dropdown:

Open vs Closed each with the appropriate click handler and displaying
the selected text or the help tip if nothing is selected.

To accomplish this, we extract a `dropdownHead` function that deals with
selected vs no selection. Since we already know whether the dropdown is
open or not, we just pass in the `Msg` type we want for the click
handler.

Because we're now handling whether an item is selected or not, the
closed view is no longer static.