Set selected item when clicked
This sets our selected item in response to a Msg sent when the user
clicks on an item. This Msg is more complex than previous ones because
it wraps a value.

This means we need to give it a value in the click handler. When
reacting to the event, we unwrap the value as part of the case
statement, just like we did for the `Maybe`.