Set search query as user types
When the user types in the text box, we set the query value on the
model, thus filtering the dropdown.

Note that the new Msg value `SearchInput` takes a paremeter, we don't
give it one in `onInput`. While `onClick` takes a Msg as its argument,
`onInput` takes a _function_ `String -> Msg` as its argument.

`SearchInput` is such a function when not given a value.

`onInput` does this because we don't know the value ahead of time.
Instead, it will put the whatever value the user has typed inside the
Msg.