Focus search input when opening dropdown
This uses a combination of `Dom.focus` and `Task.attempt` to create a
Cmd that will focus the DOM node with the given id. Since all Cmds
trigger a Msg when they are done but we just want to fire and forget, we
create a Noop Msg to deal with the situation.